revisi setengah

+++ database/migrations/2023_01_04_090805_create_product_formula_table.php
@@ -18,7 +18,7 @@ class CreateProductFormulaTable extends Migration
             $table->foreignId('merchant_id');
             $table->foreignId('product_id');
             $table->foreignId('ingredient_id');
-            $table->mediumInteger('amount');
+            $table->decimal('amount',8,2);
             $table->timestamps();
         });
     }
